Correct bug in the symptoms looked for by farming experience analysis
The contingency table was originally split according to prevalence of lameness experienced, rather than farming experience. Corrected this and redid the chi-squared analysis and the balloon plot - still no significant difference.  New chi-squared result and balloon plot is uploaded to OSF.

Added some extra annotation in the wrangling code related to the production of the contingency table too.